+++ Tools/QuickEd/Classes/UI/Preview/ScrollAreaController.h
@@ -20,21 +20,21 @@ public:
     Q_PROPERTY(QSize canvasSize READ GetCanvasSize NOTIFY CanvasSizeChanged);
     Q_PROPERTY(QSize viewSize READ GetViewSize WRITE SetViewSize NOTIFY ViewSizeChanged);
     Q_PROPERTY(QPoint position READ GetPosition WRITE SetPosition NOTIFY PositionChanged);
-    Q_PROPERTY(qreal scale READ GetScale WRITE SetScale NOTIFY ScaleChanged);
+    Q_PROPERTY(float scale READ GetScale WRITE SetScale NOTIFY ScaleChanged);
 
     ScrollAreaController(QObject* parent = nullptr);
     ~ScrollAreaController();
 
     void SetNestedControl(DAVA::UIControl* nestedControl);
     void SetMovableControl(DAVA::UIControl* movableControl);
-    void AdjustScale(qreal newScale, QPointF mousePos);
+    void AdjustScale(float newScale, QPointF mousePos);
 
     QSize GetCanvasSize() const;
     QSize GetViewSize() const;
     QPoint GetPosition() const;
-    qreal GetScale() const;
-    qreal GetMinScale() const;
-    qreal GetMaxScale() const;
+    float GetScale() const;
+    float GetMinScale() const;
+    float GetMaxScale() const;
     QPoint GetMinimumPos() const;
     QPoint GetMaximumPos() const;
 
@@ -42,13 +42,13 @@ public slots:
     void SetViewSize(QSize size);
     void SetPosition(QPoint position);
     void UpdateCanvasContentSize();
-    void SetScale(qreal scale);
+    void SetScale(float scale);
 
 signals:
     void CanvasSizeChanged(QSize canvasSize);
     void ViewSizeChanged(QSize size);
     void PositionChanged(QPoint position);
-    void ScaleChanged(qreal scale);
+    void ScaleChanged(float scale);
     void NestedControlPositionChanged(QPoint position);
 
 private:
@@ -59,9 +59,9 @@ private:
     QSize canvasSize = QSize(0, 0);
     QSize viewSize = QSize(0, 0);
     QPoint position = QPoint(0, 0);
-    qreal scale = 0.0f;
-    const qreal minScale = 0.25f;
-    const qreal maxScale = 8.0f;
+    float scale = 0.0f;
+    const float minScale = 0.25f;
+    const float maxScale = 8.0f;
     const int Margin = 50;
 };
